refactor(m2): switch persistence layer from server API to IndexedDB (PR-A A.3-A.7)
- store/projectSlice.ts: createProjectApi/deleteProjectApi -> putProject/deleteProject
- store/syncSlice.ts: updateProjectApi -> putProject (debounce/flush logic preserved)
- store/tutorialSlice.ts: /api/tutorial fetch -> tutorialRepository
- store/analysisHistorySlice.ts: /api/analysis-history fetch -> analysisHistoryRepository
- hooks/useFirestoreSync.ts -> hooks/useLocalSync.ts (rename + Dexie-backed init,
  IndexedDB unavailability surfaces a toast and isolates memory-only mode)
- App.tsx: import useLocalSync (return shape unchanged)
- components/Header.tsx: handleExportAll uses analysisHistoryRepository

historyTree remains memory-only (max 10 nodes) per ADR-0001 revision.
IndexedDB is not bound to uid (login/logout preserves local data).

Refs: docs/spec/m2/tasks.md PR-A A.3, A.4, A.5, A.6, A.7
