+++ App.tsx
@@ -17,13 +17,13 @@ import { ProjectSelectionScreen } from './components/ProjectSelectionScreen';
 import { Toast } from './components/Toast';
 import { TutorialModeSelectionModal } from './components/TutorialModeSelectionModal';
 import AppMobile from './App.mobile';
-import { useFirestoreSync } from './hooks/useFirestoreSync';
+import { useLocalSync } from './hooks/useLocalSync';
 
 import { useShallow } from 'zustand/react/shallow';
 import { EMPTY_ARRAY } from './constants';
 
 export default function App() {
-    const { isInitializing, error: syncError } = useFirestoreSync();
+    const { isInitializing, error: syncError } = useLocalSync();
     // --- Zustand Store Selectors ---
     const activeProjectId = useStore(state => state.activeProjectId);
     const allProjectsData = useStore(state => state.allProjectsData);
