+++ store/analysisHistorySlice.ts
@@ -1,5 +1,8 @@
-import { StateCreator } from 'zustand';
 import { AnalysisResult } from '../types';
+import {
+    loadAnalysisHistory as loadAnalysisHistoryFromDb,
+    saveAnalysisHistory as saveAnalysisHistoryToDb,
+} from '../db/analysisHistoryRepository';
 
 export interface AnalysisHistorySlice {
     analysisHistory: AnalysisResult[];
@@ -11,11 +14,8 @@ export const createAnalysisHistorySlice = (set, get): AnalysisHistorySlice => ({
     analysisHistory: [],
     loadAnalysisHistory: async () => {
         try {
-            const response = await fetch('/api/analysis-history');
-            if (response.ok) {
-                const data = await response.json();
-                set({ analysisHistory: data.history || [] });
-            }
+            const history = await loadAnalysisHistoryFromDb();
+            set({ analysisHistory: history });
         } catch (error) {
             console.error('Failed to load analysis history:', error);
         }
@@ -24,15 +24,11 @@ export const createAnalysisHistorySlice = (set, get): AnalysisHistorySlice => ({
         const { analysisHistory } = get();
         // Add new result to the start, limit to 2
         const updatedHistory = [newResult, ...analysisHistory].slice(0, 2);
-        
+
         set({ analysisHistory: updatedHistory });
-        
+
         try {
-            await fetch('/api/analysis-history', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ history: updatedHistory }),
-            });
+            await saveAnalysisHistoryToDb(updatedHistory);
         } catch (error) {
             console.error('Failed to save analysis history:', error);
         }
